Chore: Improve project linting, building, packages
