+++ package.json
@@ -12,7 +12,10 @@
     "bootstrap": "lerna bootstrap --use-workspaces",
     "build": "node .bin/build.js",
     "build:watch": "node .bin/build-watch.js",
-    "lint": "node .bin/test-lint.js",
+    "lint": "node .bin/lint.js",
+    "lint:esm": "node .bin/lint-esm.js",
+    "lint:pkg": "node .bin/lint-pkg.js",
+    "lint:tsc": "node .bin/lint-tsc.js",
     "prerelease": "npm run build && npm run test",
     "release": "lerna publish",
     "release:canary": "npm run prerelease && lerna publish --dist-tag canary",
@@ -27,21 +30,22 @@
     "play"
   ],
   "dependencies": {
+    "@skypack/package-check": "^0.2.2",
     "@types/react": "17.0.3",
-    "@types/react-dom": "17.0.2",
+    "@types/react-dom": "17.0.3",
     "@types/react-test-renderer": "17.0.1",
-    "@typescript-eslint/eslint-plugin": "4.18.0",
-    "@typescript-eslint/parser": "4.18.0",
+    "@typescript-eslint/eslint-plugin": "4.19.0",
+    "@typescript-eslint/parser": "4.19.0",
     "esbuild": "0.9.6",
     "eslint": "7.22.0",
     "lerna": "4.0.0",
-    "linkedom": "0.5.6",
+    "linkedom": "0.6.1",
     "magic-string": "0.25.7",
     "merge-source-map": "1.1.0",
     "nodemon": "2.0.7",
     "npm-run-all": "4.1.5",
-    "react": "17.0.1",
-    "react-test-renderer": "17.0.1",
+    "react": "17.0.2",
+    "react-test-renderer": "17.0.2",
     "terser": "5.6.1",
     "typescript": "4.2.3"
   },
