+++ packages/core/package.json
@@ -8,6 +8,11 @@
   "typings": "types/index.d.ts",
   "jsdelivr": "dist/index.iife.js",
   "unpkg": "dist/index.iife.js",
+  "exports": {
+    "require": "./dist/index.cjs",
+    "import": "./dist/index.mjs",
+    "default": "./dist/index.mjs"
+  },
   "files": [
     "dist",
     "types",
