+++ packages/stringify/types/index.d.ts
@@ -1,0 +1,3 @@
+/// <reference no-default-lib="true"/>
+
+export function stringify(value: any, replacer?: (name: string, data: any, style: any) => any): string
